third_party/blink/web_tests/platform/linux-chrome/external/wpt/storage-access-api/requestStorageAccess-dedicated-worker.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Workers inherit storage access
  assert_true: requestStorageAccess resolves without requiring a gesture. expected true got false
[FAIL] Workers don't observe parent's storage access
  assert_equals: Worker's first fetch is uncredentialed. expected "" but got "cookie=unpartitioned"
Harness: the test ran to completion.
